backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
transformers==4.57.3
tensorflow==2.20.0
tf-keras==2.20.1
torch==2.9.1
pillow==12.0.0
numpy==2.3.5
